pyproject.toml
[project]
name = "agentic-rag-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.14",
    "langchain-opendataloader-pdf>=2.0.0",
    "langchain-pinecone>=0.2.13",
    "langgraph>=1.1.8",
    "opendataloader-pdf>=2.2.1",
    "pillow>=12.2.0",
    "pymupdf>=1.27.2.2",
]
